Remove "import Syntax" from all live Haskell source files in the main build.
