+++ src/lib/RenderHtml.hs
@@ -20,12 +20,12 @@ import Control.Monad
 import Text.Megaparsec hiding (chunk)
 import Text.Megaparsec.Char as C
 
-import Paths_dex  (getDataFileName)
-import Lexing (Parser, symChar, keyWordStrs, symbol, parseit, withSource)
-import Syntax
-import PPrint
-import Serialize ()
 import Err
+import Lexing (Parser, symChar, keyWordStrs, symbol, parseit, withSource)
+import Paths_dex  (getDataFileName)
+import Serialize ()
+import Types.Misc
+import Types.Source
 
 cssSource :: T.Text
 cssSource = unsafePerformIO $
